+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## v0.23
 
 * `discrete` and `mixture` distributions now give more detail when printed.
+* Added `get_median_and_ci` to return the median and a given confidence interval for data.
 
 
 ## v0.22